Create trip Screen + hookup to api and store
Have templated out the create trip screen, setup the functionality for adding trip_name, cover image and trip dates.

then we collect locations for the trip.
  added a search for locations, on select of location we are taken to the location details screen where we collect the location name and all images attached to location. onSave of that we add the captured location details to the parent trip_form under locations. from there they are navigated back to the findLocation screen wwhere they can repeat this process until they are happy with the locations added to the current created trip. Clicking the tiles for each location will trigger the map to navigate to said location.

Validation on create Trip form, hooked up to the useTripStore(), which now houses the logic for this form and actions against a trip, ie create, fetchTrips and delete a trip.